dante - beta

+++ TournamentBuilderLib/Models/BattleResult.cs
@@ -4,7 +4,7 @@ public class BattleResult
 {
     public string Range { get; set; }
 
-    public int Result { get; set; }
+    public double Result { get; set; }
 
     public int Score { get; set; }
 }
